+++ kiro-iso-builder.py
@@ -101,6 +101,7 @@ class BuilderApp(Gtk.Application):
 
     def _on_activate(self, app):
         fn.ensure_app_dirs()
+        fn.ensure_build_conf()
         self._load_css()
         if not self.get_windows():
             BuilderWindow(app).present()